+++ data/js/platforms/dailymotion.js
@@ -71,7 +71,7 @@ websites.dailymotion = {
 			if(data.hasOwnProperty("user.screenname")){
 				if(isStreamOnline){
 					streamData.streamStatus = streamData.streamName;
-					streamData.streamGame = (data["game.title"] !== null && typeof data["game.title"] == "string")? data["game.title"] : "";
+					streamData.streamGame = (data["game.title"] != null && typeof data["game.title"] == "string")? data["game.title"] : "";
 				}
 				if(typeof data["user.avatar_720_url"] == "string" && data["user.avatar_720_url"] != ""){
 					streamData.streamOwnerLogo = data["user.avatar_720_url"];