+++ data/js/platforms/hitbox.js
@@ -24,16 +24,15 @@ websites.hitbox = {
 				let error_msg = (data.hasOwnProperty("error_msg") == true)? data.error_msg : "error";
 				return error_msg;
 			}
+			if(data.hasOwnProperty("livestream") == false){
+				return "data missing";
+			}
 			return "success";
 		},
 	"checkLiveStatus":
 		function(id, contentId, data, currentLiveStatus){
 			let streamData = currentLiveStatus;
 			if(data.hasOwnProperty("livestream") == false){
-				if(data.error_msg="no_media_found"){
-					streamData.liveStatus.API_Status = false;
-				}
-				streamData.streamName = id;
 				return null;
 			}
 			if(typeof data["livestream"][0] == "object"){
@@ -43,20 +42,20 @@ websites.hitbox = {
 				if(typeof data["category_name"] == "string" && data["category_name"] != ""){
 					streamData.streamGame = data["category_name"];
 				}
-				if(data["category_logo_large"] !== null){
+				if(data["category_logo_large"] != null){
 					streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv" + data["category_logo_large"];
-				} else if(data["category_logo_small"] !== null){
+				} else if(data["category_logo_small"] != null){
 					streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv" + data["category_logo_small"];
 				} else {
 					streamData.streamCategoryLogo = "";
 				}
-				if(streamData.streamCategoryLogo = "http://edge.sf.hitbox.tv/static/img/generic/blank.gif"){
+				if(streamData.streamCategoryLogo == "http://edge.sf.hitbox.tv/static/img/generic/blank.gif"){
 					streamData.streamCategoryLogo = "";
 				}
 				
 				if(typeof data.channel["user_logo"] == "string" && data.channel["user_logo"].indexOf("/static/img/generic/default-user-") == -1){
 					streamData.streamOwnerLogo = "http://edge.sf.hitbox.tv" + data.channel["user_logo"];
-				} else if(typeof data.channel["user_logo"] !== "string" && data.channel["user_logo"].indexOf("/static/img/generic/default-user-") == -1){
+				} else if(typeof data.channel["user_logo"] != "string" && data.channel["user_logo"].indexOf("/static/img/generic/default-user-") == -1){
 					streamData.streamOwnerLogo = "http://edge.sf.hitbox.tv" + data.channel["user_logo_small"];
 				} else {
 					streamData.streamOwnerLogo = "";
@@ -88,7 +87,7 @@ websites.hitbox = {
 				
 				if(data.following.length > 0){
 					let nextPageNumber = ((typeof pageNumber == "number")? pageNumber : 1) + 1;
-					let nextUrl = websites.hitbox.importAPI(id).url + "&offset=" + nextPageNumber;
+					let nextUrl = hitbox.importAPI(id).url + "&offset=" + nextPageNumber;
 					obj.next = {"url": nextUrl, "pageNumber": nextPageNumber};
 				} else {
 					obj.next = null;